+++ Magic/CHANGELOG.md
@@ -12,6 +12,7 @@
  - Add "current" parameter to Cancel action, to try to avoid canceling the current cast
  - Fix error loading emerald sword on startup in 1.15 and earlier
  - Add "allow_dropped_items" Selector parameter, will cancel a purchase if your inventory is full
+ - Fix for mob triggers not properly re-triggering
 
 # 7.9.8